requirements.txt
langchain==0.1.12
langchain-openai==0.0.8
streamlit==1.32.2
chromadb==0.4.24
python-dotenv==1.0.1
unstructured==0.11.8
markdown==3.5.2
tokenizers==0.15.2 